+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## 0.1.5 (WIP)
 
+* Click variation
+* Disable start button while running
+
 ## 0.1.4
 
 * Workflow file renamed from `deploy.yml` to `release.yml`